add User model
