+++ models/user.js
@@ -1,0 +1,50 @@
+const { Schema, model } = require("mongoose");
+const Joi = require("joi");
+
+const { handleMongooseError } = require("../helpers");
+
+const emailRegexp = /^\w+([\.-]?\w+)*@\w+([\.-]?\w+)*(\.\w{2,3})+$/
+
+const userSchema = new Schema({
+    name: {
+        type: String,
+        required: true
+    },
+    email: {
+        type: String,
+        match: emailRegexp,
+        unique: true,
+        required: true       
+    },
+    password: {
+        type: String,
+        minlength: 6,
+        required: true,
+    }
+}, { versionKey: false })
+
+userSchema.post("save", handleMongooseError)
+
+const registerSchema = Joi.object({
+    name: Joi.string().required(),
+    email: Joi.string().pattern(emailRegexp).required(),
+    password: Joi.string().min(6).required()
+})
+
+const loginSchema = Joi.object({
+    email: Joi.string().pattern(emailRegexp).required(),
+    password: Joi.string().min(6).required()
+})
+
+const schemas = {
+    registerSchema,
+    loginSchema
+}
+
+const User = model("user", userSchema);
+
+module.exports = {
+    User,
+    schemas,
+}
+
