+++ routes/api/auth.js
@@ -1,34 +1,7 @@
-const { Schema, model } = require("mongoose");
-const Joi = require("joi");
+const express = require("express");
 
-const { handleMongooseError } = require("../helpers");
+const router = express.Router();
 
-const emailRegexp = /^\w+([\.-]?\w+)*@\w+([\.-]?\w+)*(\.\w{2,3})+$/
-
-const userSchema = new Schema({
-    name: {
-        type: String,
-        required: true
-    },
-    email: {
-        type: String,
-        match: emailRegexp,
-        unique: true,
-        required: true       
-    },
-    password: {
-        type: String,
-        minlength: 6,
-        required: true,
-    }
-}, { versionKey: false, timestamps: true })
-
-userSchema.post("save", handleMongooseError)
-
-const registerSchema = Joi.object({
-    name: Joi.string().required(),
-    email: Joi.string().pattern(emailRegexp).required(),
-    password: Joi.string().min().required()
-})
+module.exports = router;
